Avoid i/o stalls in shell layer code
The shell packetization layer may receive data, but cannot forward it
because the next receiver's queue is already full. In this case, we need
to rely on the caller to call us again when there's room.

The simplest way to do this is to assert a "call me again later" flag.

Signed-off-by: Olaf Kirch <okir@suse.de>